scroll zoom: use fractions of power of 2 to avoid desync between key and mouse zoom

+++ client/src/proxguiqt.h
@@ -51,7 +51,7 @@ class Plot: public QWidget {
   protected:
     void paintEvent(QPaintEvent *event);
     void closeEvent(QCloseEvent *event);
-    void Zoom(float factor, int refX);
+    void Zoom(double factor, int refX);
     void Move(int offset);
     void Trim(void);
     void wheelEvent(QWheelEvent *event);
